Add the Ball Recognition using YOLOv11x, data still needs to be refined and fix large file issue and rename the file
