+++ requirements.txt
@@ -34,3 +34,4 @@ six==1.16.0
 sounddevice==0.5.1
 pygments==2.10.0
 jinja2==3.0.1
+ultralytics==8.3.78
